refactor(state): extract registry persistence seams
