+++ ci/source-architecture-budget.json
@@ -26,7 +26,7 @@
       "src/lib/runner.ts": 89,
       "src/lib/security/redact.ts": 50,
       "src/lib/state/onboard-session.ts": 34,
-      "src/lib/state/registry.ts": 102,
+      "src/lib/state/registry.ts": 99,
       "src/lib/state/state-root.ts": 23,
       "src/lib/subprocess-env.ts": 23,
       "src/lib/validation.ts": 25
